Improvements to kill_log channel

+++ commands/prune.js
@@ -1,4 +1,5 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
+const { Permissions } = require('discord.js');
 
 module.exports = {
 	data: new SlashCommandBuilder()
@@ -6,16 +7,22 @@ module.exports = {
 		.setDescription('Prune up to 99 messages.')
 		.addIntegerOption(option => option.setName('amount').setDescription('Number of messages to prune')),
 	async execute(interaction) {
-		const amount = interaction.options.getInteger('amount');
+		if (!interaction.member.permissions.has(Permissions.FLAGS.ADMINISTRATOR)) {
+			console.log('This member is NOT an administrator');
+			return interaction.reply({ content: `Sorry you need to have administrator permissions in order to complete this.`, ephemeral: true });
+		} else {		
+			const amount = interaction.options.getInteger('amount');
 
-		if (amount <= 1 || amount > 100) {
-			return interaction.reply({ content: 'You need to input a number between 1 and 99.', ephemeral: true });
+			if (amount <= 1 || amount > 100) {
+				return interaction.reply({ content: 'You need to input a number between 1 and 99.', ephemeral: true });
+			}
+			await interaction.channel.bulkDelete(amount, true).catch(error => {
+				console.error(error);
+				interaction.reply({ content: 'There was an error trying to prune messages in this channel!', ephemeral: true });
+			});
+
+			return interaction.reply({ content: `Successfully pruned \`${amount}\` messages.`, ephemeral: true });
+
 		}
-		await interaction.channel.bulkDelete(amount, true).catch(error => {
-			console.error(error);
-			interaction.reply({ content: 'There was an error trying to prune messages in this channel!', ephemeral: true });
-		});
-
-		return interaction.reply({ content: `Successfully pruned \`${amount}\` messages.`, ephemeral: true });
 	},
 };
